+++ UmbCheckout.StarterKit.Web/NotificationHandlers/OnContentDeletedNotificationHandler.cs
@@ -1,0 +1,32 @@
+﻿using Umbraco.Cms.Core.Cache;
+using Umbraco.Cms.Core.Events;
+using Umbraco.Cms.Core.Notifications;
+
+namespace UmbCheckout.StarterKit.Web.NotificationHandlers
+{
+    /// <summary>
+    /// After content is published
+    /// </summary>
+    public class OnContentDeletedNotificationHandler : INotificationHandler<ContentDeletedNotification>
+    {
+        private readonly AppCaches _appCaches;
+
+        public OnContentDeletedNotificationHandler(AppCaches appCaches)
+        {
+            _appCaches = appCaches;
+        }
+
+        public void Handle(ContentDeletedNotification notification)
+        {
+            if (notification.DeletedEntities.Any(x => x.ContentType.Alias == "productCategory"))
+            {
+                _appCaches.RuntimeCache.ClearByKey("ProductCategories");
+            }
+
+            if (notification.DeletedEntities.Any(x => x.ContentType.Alias == "blogCategory"))
+            {
+                _appCaches.RuntimeCache.ClearByKey("BlogCategories");
+            }
+        }
+    }
+}
