+++ UmbCheckout.StarterKit.Web/NotificationHandlers/OnContentMovedToRecycleBinNotification.cs
@@ -1,4 +1,5 @@
-﻿using Umbraco.Cms.Core.Cache;
+﻿using Microsoft.Extensions.Caching.Memory;
+using Umbraco.Cms.Core.Cache;
 using Umbraco.Cms.Core.Events;
 using Umbraco.Cms.Core.Notifications;
 
@@ -7,23 +8,25 @@ namespace UmbCheckout.StarterKit.Web.NotificationHandlers
     /// <summary>
     /// After content is published
     /// </summary>
-    public class OnContentDeletedNotificationHandler : INotificationHandler<ContentDeletedNotification>
+    public class OnContentMovedToRecycleBinNotification : INotificationHandler<ContentMovedToRecycleBinNotification>
     {
         private readonly AppCaches _appCaches;
 
-        public OnContentDeletedNotificationHandler(AppCaches appCaches)
+        public OnContentMovedToRecycleBinNotification(AppCaches appCaches)
         {
             _appCaches = appCaches;
         }
 
-        public void Handle(ContentDeletedNotification notification)
+        public void Handle(ContentMovedToRecycleBinNotification notification)
         {
-            if (notification.DeletedEntities.Any(x => x.ContentType.Alias == "productCategory"))
+
+            var contentItems = notification.MoveInfoCollection.Select(x => x.Entity);
+            if (contentItems.Any(x => x.ContentType.Alias == "productCategory"))
             {
                 _appCaches.RuntimeCache.ClearByKey("ProductCategories");
             }
 
-            if (notification.DeletedEntities.Any(x => x.ContentType.Alias == "blogCategory"))
+            if (contentItems.Any(x => x.ContentType.Alias == "blogCategory"))
             {
                 _appCaches.RuntimeCache.ClearByKey("BlogCategories");
             }
